Fix cleaning files from server storage

+++ validators/postValidator.js
@@ -56,7 +56,9 @@ const validatePost = [
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       // Delete uploaded file if validation fails
-      await fs.unlink(req.file.path);
+      if (req.file) {
+        await fs.unlink(req.file.path);
+      }
       return res.status(400).json({ errors: errors.array() });
     }
     next();
